Split quiz into two components
Quiz got kinda big, so I separated as much as I could
QuizDisplay just handles showing the questions to answer
QuizEndScreen simply shows the end results of the quiz
